requirements.txt
fastapi
uvicorn
python-multipart
pillow
requests
dotenv
langchain
langchain_openai